# wordflip-api/openapi.yaml
openapi: 3.0.3
info:
  title: WordFlip API
  version: 1.1.0
  description: |
    WordFlip 单词卡片学习应用 REST API（MVP）。

    **业务文档：** [requirements.md](../docs/wordflip/requirements.md) · [api-modules.md](../docs/wordflip/api-modules.md)

    **核心规则：**
    - 队列三态（unlearned/fuzzy/unknown）与稳定性权值 `stability`（S）**仅**经测验 `POST /quiz/sessions/{id}/answer` 变更
    - **skill 双轨：** `dictation` / `choice` 各一套热力 S 与 SRS；题型 `dictation`→dictation skill，`choice_en_cn`/`choice_cn_en`→choice skill
    - 组详情主展示按 `heatDisplayMode` 聚合的 `displayHeatLevel`；单 skill 快照含 `heatLevel`；三态用于今日排队与薄弱角标
    - 词书 `PUT /settings` 触发**增量追加分组**，不重建已有分组
    - 单词唯一键 `wordKey` = `en.trim().toLowerCase()`
    - SRS 间隔（天）：`INTERVALS = [1, 2, 4, 7, 15, 30]`，stage 0..5；答对后 `nextReviewAt = 用户时区当日结束 + INTERVALS[newStage]` 天（按 skill 独立）
    - **时区：** 服务端存 UTC；「当日」由 `X-Timezone` 请求头（缺省 `Asia/Shanghai`）或用户 profile 决定
    - **已掌握统计：** `stability >= 80` 且最近测验成功且建议间隔 ≥ 30 天（REQ-EBBING-7）
    - **学习完成度：** `round(masteredCount / 已入组总词数 × 100)`，已入组总词数为 0 时返回 0
servers:
  - url: /api/v1
    description: API v1
...
security:
  - bearerAuth: []
paths:
...
  /books/import/preview:
    post:
      tags: [Books]
      summary: 导入预览（解析文件）
      description: |
        支持 JSON / CSV / TXT（REQ-BOOK-6）。服务端解析后暂存结果并返回 `previewToken`（TTL 15 分钟，confirm 后作废）。
        同书内重复 `en` 去重；无效行跳过。单文件上限 5MB。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookImportPreviewResponse'
        '413':
          description: 文件过大
        '422':
          description: 解析失败或无有效单词
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 导入频率超限
  /books/import:
    post:
      tags: [Books]
      summary: 确认导入词书
      description: |
        使用 preview 阶段返回的 token 确认入库。成功后自动勾选该书（`selected=true`），并清除 `user_book_selection` 中对该书的旧勾选冲突。
        不自动追加分组，需用户再 `PUT /settings` 触发 append。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookImportConfirmRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookImportConfirmResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: previewToken 无效
        '410':
          description: previewToken 已过期
        '409':
          description: 同名 imported 词书冲突
        '429':
          description: 导入频率超限
  /books/{bookId}:
    get:
      tags: [Books]
      summary: 词书详情
      description: |
        返回单本可见词书（builtin 全员可见，或当前用户的 imported）。
        仅浏览元数据，不改变勾选与分组。
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookItem'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Books]
      summary: 删除导入词书
      description: 仅 `source=imported` 且归属当前用户。删除后从用户勾选列表移除；已入组单词保留。
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '204':
          description: 已删除
        '403':
          description: 内置词书不可删
        '404':
          $ref: '#/components/responses/NotFound'
  /books/{bookId}/words:
    get:
      tags: [Books]
      summary: 词书词条分页
      description: |
        按 `sort_order` 返回书内词条（只读浏览）。词书须对当前用户可见。
        不写入掌握度、不改分组。
      parameters:
        - $ref: '#/components/parameters/BookId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookWordsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ValidationError:
      description: 参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: object
          additionalProperties: true
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    WordSummary:
      type: object
      required: [wordKey, en, cn]
      properties:
        wordKey:
          type: string
        en:
          type: string
        cn:
          type: string
        pos:
          type: string
          nullable: true
        ph:
          type: string
          nullable: true
...
    BookItem:
      type: object
      required: [id, name, source, wordCount, selected, canDelete]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        source:
          type: string
          enum: [builtin, imported]
        wordCount:
          type: integer
        declaredCount:
          type: integer
          nullable: true
        selected:
          type: boolean
        canDelete:
          type: boolean
    BookImportPreviewResponse:
      type: object
      required: [previewToken, suggestedName, totalCount, previewWords]
      properties:
        previewToken:
          type: string
          format: uuid
          description: 服务端暂存解析结果，TTL 15 分钟
        suggestedName:
          type: string
        totalCount:
          type: integer
          description: 去重后有效词数
        deduplicatedCount:
          type: integer
          description: 同文件内被去重的重复行数
        previewWords:
          type: array
          maxItems: 6
          items:
            $ref: '#/components/schemas/WordSummary'
    BookImportConfirmRequest:
      type: object
      required: [previewToken, name]
      properties:
        previewToken:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 64
    BookImportConfirmResponse:
      type: object
      required: [book]
      properties:
        book:
          $ref: '#/components/schemas/BookItem'
          description: confirm 后 selected=true
    BookWordsResponse:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [words]
          properties:
            words:
              type: array
              items:
                $ref: '#/components/schemas/WordSummary'
...